Included tests for registration API; Included validations on the registration API;

+++ src/Zazz.Web/Models/EditUserProfileViewModel.cs
@@ -21,7 +21,7 @@ namespace Zazz.Web.Models
 
         public IEnumerable<Major> Majors { get; set; }
 
-        public byte? MajorId { get; set; }
+        public int? MajorId { get; set; }
 
         public IEnumerable<City> Cities { get; set; }
